fix double promotions - next round

+++ play_memory.cpp
@@ -188,10 +188,10 @@ void AudioPlayMemory::update(void)
 }
 
 
-#define B2M_88200 (uint32_t)((double)4294967296000.0 / AUDIO_SAMPLE_RATE_EXACT / 2.0)
-#define B2M_44100 (uint32_t)((double)4294967296000.0 / AUDIO_SAMPLE_RATE_EXACT) // 97352592
-#define B2M_22050 (uint32_t)((double)4294967296000.0 / AUDIO_SAMPLE_RATE_EXACT * 2.0)
-#define B2M_11025 (uint32_t)((double)4294967296000.0 / AUDIO_SAMPLE_RATE_EXACT * 4.0)
+#define B2M_88200 (uint32_t)((double)4294967296000.0 / (double)AUDIO_SAMPLE_RATE_EXACT / 2.0)
+#define B2M_44100 (uint32_t)((double)4294967296000.0 / (double)AUDIO_SAMPLE_RATE_EXACT) // 97352592
+#define B2M_22050 (uint32_t)((double)4294967296000.0 / (double)AUDIO_SAMPLE_RATE_EXACT * 2.0)
+#define B2M_11025 (uint32_t)((double)4294967296000.0 / (double)AUDIO_SAMPLE_RATE_EXACT * 4.0)
 
 
 uint32_t AudioPlayMemory::positionMillis(void)
